Restore error type

+++ lib/src/error.rs
@@ -2,7 +2,6 @@ use core::fmt::{self, Display, Formatter};
 use core::num::ParseIntError;
 
 /// Encoding and decoding errors.
-#[allow(variant_size_differences)]
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum Error {
     /// The input byte stream terminated prematurely.
@@ -20,9 +19,9 @@ pub enum Error {
     /// Too many digits for the color code have been encountered.
     TooManyCodeDigits {
         /// Number of digits that have been processed.
-        at_least: usize,
+        at_least: u8,
         /// Number of digits that were expected.
-        max: usize,
+        max: u8,
     },
     /// An unexpected byte has been encountered.
     UnexpectedByte(u8),
